requirements.txt
Django==1.7
snmp-passpersist
Pillow
